+++ unity-sample-environment/Assets/Scenes/sampleMultiAgent.unity
@@ -220,7 +220,7 @@ MonoBehaviour:
   url: ws://localhost:8765/ws
   cycleTimeStepSize: 0.15
   episodeTimeLength: 15
-  agent: {fileID: 1414357241}
+  agent: {fileID: 2078493469}
   environment: {fileID: 2101396359}
 --- !u!4 &370758644
 Transform:
@@ -245,13 +245,11 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: f35cce202312046d8bb7e6e255e6cc37, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  domain: localhost
-  path: ws
-  port: 8765
+  url: ws://localhost:8765/ws
   cycleTimeStepSize: 0.15
   episodeTimeLength: 15
   timeScale: 1
-  agents: []
+  agent: {fileID: 2078493469}
   environment: {fileID: 2101396359}
 --- !u!1 &644654425
 GameObject:
@@ -423,58 +421,78 @@ Transform:
   m_Children: []
   m_Father: {fileID: 644654426}
   m_RootOrder: 2
---- !u!114 &1414357241 stripped
-MonoBehaviour:
-  m_PrefabParentObject: {fileID: 11462060, guid: 8b42c37fb727d4dacbb23ed579832205,
-    type: 2}
-  m_PrefabInternal: {fileID: 1683281777}
-  m_Script: {fileID: 11500000, guid: 2b2b14c64040d407698d66027d1adbf2, type: 3}
---- !u!1001 &1683281777
+--- !u!1001 &1215490701
 Prefab:
   m_ObjectHideFlags: 0
   serializedVersion: 2
   m_Modification:
     m_TransformParent: {fileID: 1943369148}
     m_Modifications:
-    - target: {fileID: 406064, guid: 8b42c37fb727d4dacbb23ed579832205, type: 2}
+    - target: {fileID: 11469880, guid: 7807d5793a14e43459d66a7063f42614, type: 2}
+      propertyPath: rgbCameras.Array.size
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 11469880, guid: 7807d5793a14e43459d66a7063f42614, type: 2}
+      propertyPath: depthCameras.Array.size
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 483300, guid: 7807d5793a14e43459d66a7063f42614, type: 2}
       propertyPath: m_LocalPosition.x
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 406064, guid: 8b42c37fb727d4dacbb23ed579832205, type: 2}
+    - target: {fileID: 483300, guid: 7807d5793a14e43459d66a7063f42614, type: 2}
       propertyPath: m_LocalPosition.y
       value: 1
       objectReference: {fileID: 0}
-    - target: {fileID: 406064, guid: 8b42c37fb727d4dacbb23ed579832205, type: 2}
+    - target: {fileID: 483300, guid: 7807d5793a14e43459d66a7063f42614, type: 2}
       propertyPath: m_LocalPosition.z
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 406064, guid: 8b42c37fb727d4dacbb23ed579832205, type: 2}
+    - target: {fileID: 483300, guid: 7807d5793a14e43459d66a7063f42614, type: 2}
       propertyPath: m_LocalRotation.x
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 406064, guid: 8b42c37fb727d4dacbb23ed579832205, type: 2}
+    - target: {fileID: 483300, guid: 7807d5793a14e43459d66a7063f42614, type: 2}
       propertyPath: m_LocalRotation.y
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 406064, guid: 8b42c37fb727d4dacbb23ed579832205, type: 2}
+    - target: {fileID: 483300, guid: 7807d5793a14e43459d66a7063f42614, type: 2}
       propertyPath: m_LocalRotation.z
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 406064, guid: 8b42c37fb727d4dacbb23ed579832205, type: 2}
+    - target: {fileID: 483300, guid: 7807d5793a14e43459d66a7063f42614, type: 2}
       propertyPath: m_LocalRotation.w
       value: 1
       objectReference: {fileID: 0}
-    - target: {fileID: 406064, guid: 8b42c37fb727d4dacbb23ed579832205, type: 2}
+    - target: {fileID: 483300, guid: 7807d5793a14e43459d66a7063f42614, type: 2}
       propertyPath: m_RootOrder
       value: 0
       objectReference: {fileID: 0}
+    - target: {fileID: 11469880, guid: 7807d5793a14e43459d66a7063f42614, type: 2}
+      propertyPath: rgbCameras.Array.data[0]
+      value: 
+      objectReference: {fileID: 1215490704}
+    - target: {fileID: 11469880, guid: 7807d5793a14e43459d66a7063f42614, type: 2}
+      propertyPath: depthCameras.Array.data[0]
+      value: 
+      objectReference: {fileID: 1215490703}
     m_RemovedComponents: []
-  m_ParentPrefab: {fileID: 100100000, guid: 8b42c37fb727d4dacbb23ed579832205, type: 2}
+  m_ParentPrefab: {fileID: 100100000, guid: 7807d5793a14e43459d66a7063f42614, type: 2}
   m_IsPrefabParent: 0
---- !u!4 &1683281778 stripped
+--- !u!4 &1215490702 stripped
 Transform:
-  m_PrefabParentObject: {fileID: 406064, guid: 8b42c37fb727d4dacbb23ed579832205, type: 2}
-  m_PrefabInternal: {fileID: 1683281777}
+  m_PrefabParentObject: {fileID: 483300, guid: 7807d5793a14e43459d66a7063f42614, type: 2}
+  m_PrefabInternal: {fileID: 1215490701}
+--- !u!20 &1215490703 stripped
+Camera:
+  m_PrefabParentObject: {fileID: 2009838, guid: 7807d5793a14e43459d66a7063f42614,
+    type: 2}
+  m_PrefabInternal: {fileID: 1215490701}
+--- !u!20 &1215490704 stripped
+Camera:
+  m_PrefabParentObject: {fileID: 2015576, guid: 7807d5793a14e43459d66a7063f42614,
+    type: 2}
+  m_PrefabInternal: {fileID: 1215490701}
 --- !u!1001 &1796151572
 Prefab:
   m_ObjectHideFlags: 0
@@ -550,13 +568,19 @@ Transform:
   m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children:
-  - {fileID: 1683281778}
+  - {fileID: 1215490702}
   m_Father: {fileID: 0}
   m_RootOrder: 1
 --- !u!4 &1950425914 stripped
 Transform:
   m_PrefabParentObject: {fileID: 445476, guid: 4ecd053992cf34b58866b793d44e5a99, type: 2}
   m_PrefabInternal: {fileID: 182423480}
+--- !u!114 &2078493469 stripped
+MonoBehaviour:
+  m_PrefabParentObject: {fileID: 11469880, guid: 7807d5793a14e43459d66a7063f42614,
+    type: 2}
+  m_PrefabInternal: {fileID: 1215490701}
+  m_Script: {fileID: 11500000, guid: 2b2b14c64040d407698d66027d1adbf2, type: 3}
 --- !u!1 &2101396357
 GameObject:
   m_ObjectHideFlags: 0
