+++ unity-sample-environment/Assets/Scenes/sampleMultiAgent.unity.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 763517e1e00ea472da8ef9ba2e4f074d
+timeCreated: 1461161446
+licenseType: Pro
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
