+++ unity-sample-environment/Assets/Scripts/AIClient.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: ff84539c6ea394688af2bcd1697c87e7
+timeCreated: 1461164401
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
